answering-service/requirements.txt
pika
openai
python-dotenv
requests